add debug build

+++ .jenkinsci/builders/x64-linux-release-build-steps.groovy
@@ -1,13 +1,8 @@
 #!/usr/bin/env groovy
 
-def releaseDockerManifestPush(dockerImageObj, dockerTag, environment) {
+def releaseDockerManifestPush(dockerImageObj, String dockerTag, List environment) {
   manifest = load ".jenkinsci/utils/docker-manifest.groovy"
-  platform = sh(script: 'uname -m', returnStdout: true).trim()
   withEnv(environment) {
-    docker.withRegistry('https://registry.hub.docker.com', 'docker-hub-credentials') {
-      sh 'echo "Pushing ${platform}-${dockerTag}"'
-      //dockerImageObj.push("${platform}-${dockerTag}")
-    }
     if (manifest.manifestSupportEnabled()) {
       manifest.manifestCreate("${env.DOCKER_REGISTRY_BASENAME}:${dockerTag}",
         ["${env.DOCKER_REGISTRY_BASENAME}:x86_64-${dockerTag}",
@@ -26,10 +21,14 @@ def releaseDockerManifestPush(dockerImageObj, dockerTag, environment) {
         manifest.manifestPush("${env.DOCKER_REGISTRY_BASENAME}:${dockerTag}", login, password)
       }
     }
+    else {
+      sh('echo [WARNING] Docker CLI does not support manifest management features. Manifest will not be updated')
+    }
   }
 }
 
-def buildSteps(nodeLabel, parallelism, compilerVersion, dockerTag, environment) {
+def buildSteps(String nodeLabel, int parallelism, String compilerVersion, String dockerTag,
+      boolean pushDockerTag, List environment) {
   node (nodeLabel) {
     stage('Build') {
       withEnv(environment) {
@@ -43,19 +42,50 @@ def buildSteps(nodeLabel, parallelism, compilerVersion, dockerTag, environment) 
         iC.pull()
         iC.inside("-v /var/jenkins/ccache:${env.CCACHE_RELEASE_DIR}") {
           sh "curl -L -o build/iroha-fff.deb http://de.archive.ubuntu.com/ubuntu/pool/main/n/netkit-telnet/telnet_0.17-40_amd64.deb"
+          sh "tar -zcf build/iroha.tar.gz build/iroha-fff.deb"
           // build.cmakeConfigure("-DCMAKE_CXX_COMPILER=${compilers[compilerVersion]['cxx_compiler']} \
           //   -DCMAKE_CC_COMPILER=${compilers[compilerVersion]['cc_compiler']} -DCMAKE_BUILD_TYPE=Release \
           //   -DPACKAGE_DEB=ON -DPACKAGE_TGZ=ON -DCOVERAGE=OFF -DTESTING=OFF")
           // build.cmakeBuild("--target package", parallelism)
+          sh "mv ./build/iroha-*.deb ./build/iroha.deb"
         }
-        sh "mv ./build/iroha-*.deb docker/release/iroha.deb"
-        iCRelease = docker.build("${env.DOCKER_REGISTRY_BASENAME}:${scmVars.GIT_COMMIT}-${env.BUILD_NUMBER}-release", "--no-cache -f docker/release/Dockerfile ${WORKSPACE}/docker/release")
-        releaseDockerManifestPush(iCRelease, dockerTag, environment)
-        sh "docker rmi ${iCRelease.id}"
-        // sh "echo qqqqRunning on ${nodeLabel}"
-        // sh "sleep 20"
-        // sh "sleep 1"
+        if (pushDockerTag) {
+          sh "cp ./build/iroha.deb docker/release/iroha.deb"
+          iCRelease = docker.build("${env.DOCKER_REGISTRY_BASENAME}:${scmVars.GIT_COMMIT}-${env.BUILD_NUMBER}-release", "--no-cache -f docker/release/Dockerfile ${WORKSPACE}/docker/release")
+          utils.dockerPush(iC, "${platform}-${dockerTag}")
+          releaseDockerManifestPush(iCRelease, dockerTag, environment)
+          sh "docker rmi ${iCRelease.id}"
+        }
       }
+    }
+  }
+}
+
+def alwaysPostSteps(List environment) {
+  withEnv(environment) {
+    //cleanWs()
+  }
+}
+
+def successPostSteps(scmVars, List environment) {
+  withEnv(environment) {
+    artifacts = load ".jenkinsci/artifacts.groovy"
+    filesToUpload = []
+    platform = sh(script: 'uname -m', returnStdout: true).trim()
+    sh("mkdir build/artifacts")
+    sh("mv ./build/iroha.deb ./build/iroha.tar.gz build/artifacts")
+    filePaths = [ './build/artifacts/iroha.deb', './build/artifacts/iroha.tar.gz' ]
+    filePaths.each {
+      filesToUpload.add("${it}")
+      filesToUpload.add(artifacts.writeStringIntoFile(artifacts.md5sumLinux("${it}"), "${it}.md5sum"))
+      filesToUpload.add(artifacts.writeStringIntoFile(artifacts.sha256sum("${it}"), "${it}.sha256sum"))
+      filesToUpload.add(artifacts.gpgDetachedSignatureLinux("${it}", "${it}.ascfile",'ci_gpg_privkey', 'ci_gpg_masterkey'))
+    }
+    uploadPath = sprintf('/iroha/linux/%1$s/%2$s-%3$s-%4$s',
+        [platform, scmVars.GIT_LOCAL_BRANCH, sh(script: 'date "+%Y%m%d"', returnStdout: true).trim(),
+        scmVars.GIT_COMMIT.substring(0,6)])
+    filesToUpload.each {
+      artifacts.fileUploadWithCredentials("${it}", 'ci_nexus', "https://nexus.iroha.tech/repository/artifacts/${uploadPath}")
     }
   }
 }
